Fix closingParens getting too many values
We're only interested in two closingParens values: first and the last one.
We append to this list only so that we can get the first value and once we consume the first value we want to replace it with the last.
Instead of appending to closingParens on every RIGHT_PAREN only append the first and last value seen.
This fixes the

(foo * on() bar) and on(job) sum(foo) by(abc) > 20

test case.

Signed-off-by: Lukasz Mierzwa <l.mierzwa@gmail.com>