bug: Add a unit test to show a bug in WriteChunks
WriteChunks uses the maximum size of the chunk length varint field
instead of the actual size of the varint field. This way, the chunk size
is overestimated by up to 4 bytes. This estimation error accumulates in
WriteChunks such that a new segment file might be created too early,
while the next chunk could actually fit in the current segment file.

The unit test shows, in this example, there is an estimation error of 8
bytes. If the estimationError variable is set to 0, WriteChunks creates
two segment files instead of one.

In compact.go, in PopulateBlock, WriteChunks is called for each series
so the estimation error accumulates only for a single series. Hence, the
impact is small. Nevertheless, the correct approach would be to use the
actual size of the varint field.

In the following commits, this bug will be fixed so that we can call
ChunkSizeInBytes.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Signed-off-by: Victor Herrero Otal <victor.herrero.otal@sap.com>