bugfix: Use actual size of the chunk length varint in WriteChunks
Previously, WriteChunks used the maximum size of data length varint
field for the chunk size estimation. Now, it uses the actual size.

The unit test shows the segment file can be filled with chunks up to its
maximum size.

We think the performance overhead of calling PutUvarint with a shared
array is negligible.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Signed-off-by: Victor Herrero Otal <victor.herrero.otal@sap.com>