tsdb/test: Demonstrate a size estimation bug in WriteChunks
Add a unit test that reveals a bug in WriteChunks: WriteChunks
overestimates the size of a chunk due to using the maximum varint length
instead of the actual varint length for the chunk size field. This can
cause segment files to be split unnecessarily, as the estimation error
accumulates.

The test shows that the Writer can create a single segment containing
two chunks given a large enough segment file size limit, and the
resulting segment file size is 'n'. The second half of the test shows
that, given a segment size limit of 'n', the writer unexpectedly splits
the chunks into two segments instead of one. The Writer needs a segment
file size limit a few bytes larger than the actual file size to write
both chunks in a single segment.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Signed-off-by: Victor Herrero Otal <victor.herrero.otal@sap.com>