tsdb: Use actual size of chunk length varint in WriteChunks
Fix a bug where WriteChunks used the maximum possible size for the chunk
length varint field when estimating chunk sizes, which could lead to
overestimation and unnecessary segment splits. Now, it uses the actual
size of the varint for each chunk.

This bugfix ensures segment files are filled more accurately, as
demonstrated by the updated unit test. The performance overhead of
calling PutUvarint with a shared buffer is considered negligible.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Signed-off-by: Victor Herrero Otal <victor.herrero.otal@sap.com>